wallet: split itest account helpers
